+++ mayan/apps/authentication/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url
 from django.urls import path
 
 from .views.authentication_views import (
-    MayanLoginView, MayanLogoutView, MayanPasswordChangeDoneView,
+    MayanMultiStepLoginView, MayanLogoutView, MayanPasswordChangeDoneView,
     MayanPasswordChangeView, MayanPasswordResetCompleteView,
     MayanPasswordResetConfirmView, MayanPasswordResetDoneView,
     MayanPasswordResetView, UserSetPasswordView
@@ -12,8 +12,12 @@ from .views.impersonation_views import (
     UserImpersonateStartView
 )
 
+
 urlpatterns_authenticattion = [
-    url(regex=r'^login/$', view=MayanLoginView.as_view(), name='login_view'),
+    url(
+        regex=r'^login/$', name='login_view',
+        view=MayanMultiStepLoginView.as_view()
+    ),
     url(
         regex=r'^logout/$', view=MayanLogoutView.as_view(), name='logout_view'
     ),