+++ mayan/apps/authentication/views/authentication_views.py
@@ -1,4 +1,7 @@
+from collections import OrderedDict
+
 from django.contrib import messages
+from django.contrib.auth import login as django_auth_login
 from django.contrib.auth.forms import SetPasswordForm
 from django.contrib.auth.views import (
     LoginView, LogoutView, PasswordChangeDoneView, PasswordChangeView,
@@ -6,11 +9,14 @@ from django.contrib.auth.views import (
     PasswordResetDoneView, PasswordResetView
 )
 from django.core.exceptions import PermissionDenied
+from django.forms import formsets
 from django.http import HttpResponseRedirect
 from django.shortcuts import redirect
 from django.urls import reverse, reverse_lazy
+from django.utils.decorators import classonlymethod
 from django.utils.translation import ungettext, ugettext_lazy as _
 
+from formtools.wizard.views import SessionWizardView
 from stronghold.views import StrongholdPublicMixin
 
 import mayan
@@ -21,48 +27,111 @@ from mayan.apps.user_management.permissions import permission_user_edit
 from mayan.apps.user_management.querysets import get_user_queryset
 from mayan.apps.views.generics import MultipleObjectFormActionView
 
-from ..forms import EmailAuthenticationForm, UsernameAuthenticationForm
-from ..settings import (
-    setting_disable_password_reset, setting_login_method,
-    setting_maximum_session_length
-)
+from ..classes import AuthenticationBackend
+from ..forms import AuthenticationFormBase
+from ..settings import setting_disable_password_reset
 
 
-class MayanLoginView(StrongholdPublicMixin, LoginView):
-    """
-    Control how the use is to be authenticated, options are 'email' and
-    'username'
-    """
+class MayanMultiStepLoginView(
+    StrongholdPublicMixin, SessionWizardView, LoginView
+):
     extra_context = {
         'appearance_type': 'plain'
     }
-    template_name = 'authentication/login.html'
     redirect_authenticated_user = True
+    template_name = 'authentication/login.html'
 
-    def form_valid(self, form):
-        result = super().form_valid(form=form)
-        remember_me = form.cleaned_data.get('remember_me')
+    @staticmethod
+    def form_list_property(self):
+        """
+        Return the processed form list after the view has initialized.
+        """
+        self.authentication_backend = AuthenticationBackend.cls_get_instance()
 
-        # remember_me values:
-        # True - long session
-        # False - short session
-        # None - Form has no remember_me value and we let the session
-        # expiration default.
+        form_list = self.authentication_backend.get_form_list()
 
-        if remember_me is True:
-            self.request.session.set_expiry(
-                setting_maximum_session_length.value
-            )
-        elif remember_me is False:
-            self.request.session.set_expiry(0)
+        computed_form_list = OrderedDict()
+
+        for form_index, form in enumerate(iterable=form_list):
+            computed_form_list[str(form_index)] = form
+
+        for form in computed_form_list.values():
+            if issubclass(form, formsets.BaseFormSet):
+                form = form.form
+
+        return computed_form_list
+
+    @classonlymethod
+    def as_view(cls, *args, **kwargs):
+        # SessionWizardView needs at least one form in order to be
+        # initialized as a view. Declare one empty form and then change the
+        # form list in the .dispatch() method.
+        class EmptyForm(AuthenticationFormBase):
+            """Empty form"""
+
+        cls.form_list = [EmptyForm]
+
+        # Allow super to initialize and pass the form_list len() assert
+        # before replacing the form_list attribute with our property.
+        result = super().as_view(*args, **kwargs)
+
+        def null_setter(self, value):
+            return
+
+        cls.form_list = property(
+            fget=MayanMultiStepLoginView.form_list_property,
+            fset=null_setter
+        )
 
         return result
 
-    def get_form_class(self):
-        if setting_login_method.value == 'email':
-            return EmailAuthenticationForm
+    def get_context_data(self, form, **kwargs):
+        context = super().get_context_data(form=form, **kwargs)
+
+        wizard_step = self.form_list[self.steps.current]
+
+        if self.storage.current_step == self.steps.last:
+            context['submit_label'] = None
         else:
-            return UsernameAuthenticationForm
+            context['submit_label'] = _('Next')
+
+        context.update(
+            {
+                'form_css_classes': 'form-hotkey-double-click',
+                'step_title': _(
+                    'Step %(step)d of %(total_steps)d: %(step_label)s'
+                ) % {
+                    'step': self.steps.step1, 'total_steps': len(self.form_list),
+                    'step_label': wizard_step._label
+                },
+                'wizard_step': wizard_step,
+                'wizard_steps': self.form_list
+            }
+        )
+        return context
+
+    def get_form_kwargs(self, step):
+        kwargs = super().get_form_kwargs()
+        kwargs['request'] = self.request
+        kwargs['wizard'] = self
+        return kwargs
+
+    def done(self, form_list, **kwargs):
+        """
+        Perform the same function as Django's .form_valid()
+        """
+        kwargs = self.get_all_cleaned_data()
+        self.authentication_backend.process(
+            form_list=form_list, request=self.request,
+            kwargs=kwargs
+        )
+        user = self.authentication_backend.identify(
+            form_list=form_list, request=self.request,
+            kwargs=kwargs
+        )
+
+        django_auth_login(request=self.request, user=user)
+        return HttpResponseRedirect(redirect_to=self.get_success_url())
 
 
 class MayanLogoutView(LogoutView):
@@ -97,14 +166,18 @@ class MayanPasswordChangeView(PasswordChangeView):
         return super().dispatch(*args, **kwargs)
 
 
-class MayanPasswordResetCompleteView(StrongholdPublicMixin, PasswordResetCompleteView):
+class MayanPasswordResetCompleteView(
+    StrongholdPublicMixin, PasswordResetCompleteView
+):
     extra_context = {
         'appearance_type': 'plain'
     }
     template_name = 'authentication/password_reset_complete.html'
 
 
-class MayanPasswordResetConfirmView(StrongholdPublicMixin, PasswordResetConfirmView):
+class MayanPasswordResetConfirmView(
+    StrongholdPublicMixin, PasswordResetConfirmView
+):
     extra_context = {
         'appearance_type': 'plain'
     }
@@ -154,7 +227,9 @@ class UserSetPasswordView(MultipleObjectFormActionView):
     object_permission = permission_user_edit
     pk_url_kwarg = 'user_id'
     source_queryset = get_user_queryset()
-    success_message = _('Password change request performed on %(count)d user')
+    success_message = _(
+        'Password change request performed on %(count)d user'
+    )
     success_message_plural = _(
         'Password change request performed on %(count)d users'
     )
@@ -174,7 +249,9 @@ class UserSetPasswordView(MultipleObjectFormActionView):
             result.update(
                 {
                     'object': queryset.first(),
-                    'title': _('Change password for user: %s') % queryset.first()
+                    'title': _(
+                        'Change password for user: %s'
+                    ) % queryset.first()
                 }
             )