+++ HISTORY.rst
@@ -120,6 +120,28 @@
   - Add a regular expression parser to replace values and a regular
     expression validator.
 
+- Authentication refactor:
+
+  - Subclass Django's authentication views to add multi form and multi factor
+    authentication.
+  - Add support for authentication backends. Authentication backends are
+    able to control and customize the entire login process, including
+    the forms presented to the user. Authentication backends can use mixins
+    and can be subclassed to mix and expand their capabilities.
+    Included authentication mixins: ``AuthenticationBackendRememberMeMixin``
+    Included authentication backends:
+    ``AuthenticationBackendModelUsernamePassword``,
+    ``AuthenticationBackendModelEmailPassword``.
+    Apps define authentication backends in the module
+    ``authentication_backends.py``.
+  - Removed the now unused ``EmailAuthBackend`` class.
+  - New settings:
+
+    - ``AUTHENTICATION_BACKEND`` which must be the dotted path
+      to the backend used to process user authentication.
+    - ``AUTHENTICATION_BACKEND_ARGUMENTS`` which is an optional YAML
+      structure to pass to the authentication backend.
+
 4.1.4 (2021-12-01)
 ==================
 - Changes merged from versions 4.0.20 and 4.0.21.